requirements.txt
fastapi==0.127.0
uvicorn==0.40.0
sqlalchemy==2.0.45
psycopg2-binary
pydantic==2.12.5
pydantic[email]
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
bcrypt==3.2.0
python-dotenv==1.2.1
python-multipart
alembic
openpyxl
reportlab
pandas
slowapi
